add router and new routes
