+++ .gitignore
@@ -4,3 +4,4 @@ coverage/
 node_modules/
 .yarn.lock
 package-lock.json
+src/lib/index.js
