+++ src/main.js
@@ -1,20 +1,6 @@
-/* eslint-disable import/no-cycle */
-import { bienvenida } from './components/Bienvenida.js';
-import { register } from './components/Register.js';
-import { login } from './components/Login.js';
-import { ForgotPassword } from './components/ForgotPassword.js';
-import { MainLogin } from './components/MainLogin.js';
+import { routes } from './components/Router.js';
 
 const rootDiv = document.getElementById('root');
-
-const routes = {
-  '/': bienvenida,
-  '/Register': register,
-  '/Login': login,
-  '/ForgotPasword': ForgotPassword,
-  '/MainLogin': MainLogin,
-
-};
 
 export const onNavigate = (pathname) => {
   window.history.pushState(
